+++ src/main/java/org/folio/qm/service/impl/LinksServiceImpl.java
@@ -1,12 +1,16 @@
 package org.folio.qm.service.impl;
 
+import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.folio.qm.client.LinksClient;
 import org.folio.qm.client.LinksClient.InstanceLink;
 import org.folio.qm.client.LinksClient.InstanceLinks;
+import org.folio.qm.domain.dto.BaseMarcRecord;
 import org.folio.qm.domain.dto.FieldItem;
+import org.folio.qm.domain.dto.LinkDetails;
 import org.folio.qm.domain.dto.MarcFormat;
-import org.folio.qm.domain.dto.QuickMarc;
+import org.folio.qm.domain.dto.QuickMarcEdit;
+import org.folio.qm.domain.dto.QuickMarcView;
 import org.folio.qm.service.LinkingRulesService;
 import org.folio.qm.service.LinksService;
 import org.springframework.stereotype.Service;
@@ -19,7 +23,7 @@ public class LinksServiceImpl implements LinksService {
   private final LinkingRulesService linkingRulesService;
 
   @Override
-  public void setRecordLinks(QuickMarc qmRecord) {
+  public void setRecordLinks(QuickMarcView qmRecord) {
     if (!verifyFormat(qmRecord)) {
       return;
     }
@@ -29,7 +33,7 @@ public class LinksServiceImpl implements LinksService {
   }
 
   @Override
-  public void updateRecordLinks(QuickMarc qmRecord) {
+  public void updateRecordLinks(QuickMarcEdit qmRecord) {
     if (!verifyFormat(qmRecord)) {
       return;
     }
@@ -38,24 +42,24 @@ public class LinksServiceImpl implements LinksService {
     linksClient.putLinksByInstanceId(qmRecord.getExternalId(), instanceLinks);
   }
 
-  private boolean verifyFormat(QuickMarc quickMarc) {
+  private boolean verifyFormat(BaseMarcRecord quickMarc) {
     return MarcFormat.BIBLIOGRAPHIC.equals(quickMarc.getMarcFormat());
   }
 
-  private InstanceLinks extractLinks(QuickMarc quickMarc) {
+  private InstanceLinks extractLinks(QuickMarcEdit quickMarc) {
     var links = quickMarc.getFields().stream()
-      .filter(fieldItem -> fieldItem.getAuthorityId() != null)
+      .filter(fieldItem -> fieldItem.getLinkDetails() != null)
       .map(fieldItem -> new InstanceLink()
         .setInstanceId(quickMarc.getExternalId())
-        .setAuthorityId(fieldItem.getAuthorityId())
-        .setAuthorityNaturalId(fieldItem.getAuthorityNaturalId())
-        .setLinkingRuleId(fieldItem.getLinkingRuleId()))
+        .setAuthorityId(fieldItem.getLinkDetails().getAuthorityId())
+        .setAuthorityNaturalId(fieldItem.getLinkDetails().getAuthorityNaturalId())
+        .setLinkingRuleId(fieldItem.getLinkDetails().getLinkingRuleId()))
       .toList();
 
     return new InstanceLinks(links, links.size());
   }
 
-  private void populateLinks(QuickMarc qmRecord, InstanceLinks instanceLinks) {
+  private void populateLinks(QuickMarcView qmRecord, InstanceLinks instanceLinks) {
     var linkingRules = linkingRulesService.getLinkingRules();
     instanceLinks.getLinks().forEach(instanceLink ->
       linkingRules.stream()
@@ -68,7 +72,10 @@ public class LinksServiceImpl implements LinksService {
             populateLink(fields.get(0), instanceLink);
           } else {
             fields.stream()
-              .filter(fieldItem -> instanceLink.getAuthorityId().equals(fieldItem.getAuthorityId()))
+              .filter(fieldItem -> instanceLink.getAuthorityId().equals(
+                Optional.ofNullable(fieldItem.getLinkDetails())
+                  .map(LinkDetails::getAuthorityId)
+                  .orElse(null)))
               .forEach(fieldItem -> populateLink(fieldItem, instanceLink));
           }
         })
@@ -76,8 +83,11 @@ public class LinksServiceImpl implements LinksService {
   }
 
   private void populateLink(FieldItem fieldItem, InstanceLink instanceLink) {
-    fieldItem.setAuthorityId(instanceLink.getAuthorityId());
-    fieldItem.setAuthorityNaturalId(instanceLink.getAuthorityNaturalId());
-    fieldItem.setLinkingRuleId(instanceLink.getLinkingRuleId());
+    var linkDetails = fieldItem.getLinkDetails();
+    linkDetails.setAuthorityId(instanceLink.getAuthorityId());
+    linkDetails.setAuthorityNaturalId(instanceLink.getAuthorityNaturalId());
+    linkDetails.setLinkingRuleId(instanceLink.getLinkingRuleId());
+    linkDetails.setStatus(instanceLink.getStatus());
+    linkDetails.setErrorCause(instanceLink.getErrorCause());
   }
 }