+++ src/test/java/org/folio/qm/service/impl/LinksServiceImplTest.java
@@ -1,7 +1,10 @@
 package org.folio.qm.service.impl;
 
 import static java.util.Collections.singletonList;
+import static org.apache.commons.lang3.StringUtils.EMPTY;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.folio.qm.support.utils.testentities.TestEntitiesUtils.LINK_ERROR_CAUSE;
+import static org.folio.qm.support.utils.testentities.TestEntitiesUtils.LINK_STATUS_ERROR;
 import static org.junit.jupiter.params.provider.Arguments.arguments;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.verify;
@@ -20,8 +23,10 @@ import org.folio.qm.client.LinksClient;
 import org.folio.qm.client.LinksClient.InstanceLink;
 import org.folio.qm.client.LinksClient.InstanceLinks;
 import org.folio.qm.domain.dto.FieldItem;
+import org.folio.qm.domain.dto.LinkDetails;
 import org.folio.qm.domain.dto.MarcFormat;
-import org.folio.qm.domain.dto.QuickMarc;
+import org.folio.qm.domain.dto.QuickMarcEdit;
+import org.folio.qm.domain.dto.QuickMarcView;
 import org.folio.qm.support.types.UnitTest;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -39,6 +44,7 @@ class LinksServiceImplTest {
   private static final String AUTHORITY_ID = "b9a5f035-de63-4e2c-92c2-07240c88b817";
   private static final int LINKING_RULE_ID = 1;
   private static final String BIB_TAG = "650";
+  private static final String LINK_STATUS_ACTUAL = "ACTUAL";
 
   @Mock
   private LinksClient linksClient;
@@ -81,8 +87,63 @@ class LinksServiceImplTest {
           getFieldItem(),
           getFieldItem()),
         0
+      ),
+      arguments(
+        List.of(
+          getInstanceLink(UUID.fromString(AUTHORITY_ID), LINK_STATUS_ERROR, LINK_ERROR_CAUSE)),
+        List.of(
+          getFieldItem(AUTHORITY_ID)),
+        1
       )
     );
+  }
+
+  public static Stream<Arguments> updateLinksTestData() {
+    return Stream.of(
+      arguments(List.of(
+        getFieldItem(),
+        getFieldItemLinked()
+      ), 1),
+      arguments(singletonList(
+        getFieldItemLinkedWithError()
+      ), 1),
+      arguments(Collections.emptyList(), 0)
+    );
+  }
+
+  private static FieldItem getFieldItem() {
+    return new FieldItem().tag(BIB_TAG).indicators(List.of("\\", "\\")).content("$a bcdefghijklmn");
+  }
+
+  private static FieldItem getFieldItem(String authorityId) {
+    return getFieldItem()
+      .linkDetails(new LinkDetails().authorityId(UUID.fromString(authorityId)).authorityNaturalId("12345"));
+  }
+
+  private static FieldItem getFieldItemLinked() {
+    var fieldItem = getFieldItem(AUTHORITY_ID);
+    fieldItem.getLinkDetails().linkingRuleId(LINKING_RULE_ID);
+    return fieldItem;
+  }
+
+  private static FieldItem getFieldItemLinkedWithError() {
+    var fieldItem = getFieldItem(AUTHORITY_ID);
+    fieldItem.getLinkDetails().linkingRuleId(LINKING_RULE_ID).status(LINK_STATUS_ERROR).errorCause(LINK_ERROR_CAUSE);
+    return fieldItem;
+  }
+
+  private static InstanceLink getInstanceLink(UUID authorityId, String status, String errorCause) {
+    return new InstanceLink(RandomUtils.nextInt(),
+      authorityId,
+      "12345",
+      UUID.fromString("b9a5f035-de63-4e2c-92c2-07240c89b817"),
+      LINKING_RULE_ID,
+      status,
+      errorCause);
+  }
+
+  private static InstanceLink getInstanceLink(UUID authorityId) {
+    return getInstanceLink(authorityId, LINK_STATUS_ACTUAL, EMPTY);
   }
 
   @ParameterizedTest
@@ -96,51 +157,44 @@ class LinksServiceImplTest {
     when(linksClient.fetchLinksByInstanceId(any())).thenReturn(Optional.of(instanceLinks));
     when(rulesService.getLinkingRules()).thenReturn(linkingRules);
 
-    var record = getQuickMarc(fieldItemsMock);
+    var record = getQuickMarcView(fieldItemsMock);
     service.setRecordLinks(record);
 
     var linkedFields = record.getFields().stream()
-      .filter(fieldItem -> fieldItem.getAuthorityId() != null)
+      .filter(fieldItem -> fieldItem.getLinkDetails() != null)
       .collect(Collectors.toList());
     assertThat(linkedFields)
       .hasSize(expectedLinkedFieldsCount);
 
     for (int i = 0; i < linkedFields.size(); i++) {
-      var linkedField = linkedFields.get(i);
-      assertThat(linkedField.getAuthorityId())
+      var linkDetails = linkedFields.get(i).getLinkDetails();
+      assertThat(linkDetails.getAuthorityId())
         .isEqualTo(links.get(i).getAuthorityId());
-      assertThat(linkedField.getAuthorityNaturalId())
+      assertThat(linkDetails.getAuthorityNaturalId())
         .isEqualTo(links.get(i).getAuthorityNaturalId());
-      assertThat(linkedField.getLinkingRuleId())
+      assertThat(linkDetails.getLinkingRuleId())
         .isEqualTo(links.get(i).getLinkingRuleId());
+      assertThat(linkDetails.getStatus())
+        .isEqualTo(links.get(i).getStatus());
+      assertThat(linkDetails.getErrorCause())
+        .isEqualTo(links.get(i).getErrorCause());
     }
-  }
-
-
-  public static Stream<Arguments> updateLinksTestData() {
-    return Stream.of(
-      arguments(List.of(
-        getFieldItem(),
-        getFieldItemLinked()
-      ), 1),
-      arguments(Collections.emptyList(), 0)
-    );
   }
 
   @ParameterizedTest
   @MethodSource("updateLinksTestData")
   void testRecordLinksUpdated(List<FieldItem> fieldsMock, Integer expectedLinkUpdates) {
-    var quickMarcMock = getQuickMarc(fieldsMock);
+    var quickMarcMock = getQuickMarcEdit(fieldsMock);
 
     service.updateRecordLinks(quickMarcMock);
 
     var expectedLinks = fieldsMock.stream()
-      .filter(fieldItem -> fieldItem.getAuthorityId() != null)
+      .filter(fieldItem -> fieldItem.getLinkDetails() != null)
       .map(fieldItem -> new InstanceLink()
         .setInstanceId(quickMarcMock.getExternalId())
-        .setAuthorityId(fieldItem.getAuthorityId())
-        .setAuthorityNaturalId(fieldItem.getAuthorityNaturalId())
-        .setLinkingRuleId(fieldItem.getLinkingRuleId()))
+        .setAuthorityId(fieldItem.getLinkDetails().getAuthorityId())
+        .setAuthorityNaturalId(fieldItem.getLinkDetails().getAuthorityNaturalId())
+        .setLinkingRuleId(fieldItem.getLinkDetails().getLinkingRuleId()))
       .collect(Collectors.toList());
     var expectedInstanceLinks = new InstanceLinks(expectedLinks, expectedLinks.size());
 
@@ -151,7 +205,7 @@ class LinksServiceImplTest {
   @ParameterizedTest
   @EnumSource(value = MarcFormat.class, names = "BIBLIOGRAPHIC", mode = EnumSource.Mode.EXCLUDE)
   void testRecordLinksSetNotBib(MarcFormat marcFormat) {
-    var quickMarc = new QuickMarc().marcFormat(marcFormat);
+    var quickMarc = new QuickMarcView().marcFormat(marcFormat);
     service.setRecordLinks(quickMarc);
     verifyNoInteractions(linksClient);
   }
@@ -159,33 +213,19 @@ class LinksServiceImplTest {
   @ParameterizedTest
   @EnumSource(value = MarcFormat.class, names = "BIBLIOGRAPHIC", mode = EnumSource.Mode.EXCLUDE)
   void testRecordLinksUpdateNotBib(MarcFormat marcFormat) {
-    var quickMarc = new QuickMarc().marcFormat(marcFormat);
+    var quickMarc = new QuickMarcEdit().marcFormat(marcFormat);
     service.updateRecordLinks(quickMarc);
     verifyNoInteractions(linksClient);
   }
 
-  private static FieldItem getFieldItem() {
-    return new FieldItem().tag(BIB_TAG).indicators(List.of("\\", "\\")).content("$a bcdefghijklmn");
+  private QuickMarcView getQuickMarcView(List<FieldItem> fieldItems) {
+    return new QuickMarcView().marcFormat(MarcFormat.BIBLIOGRAPHIC)
+      .externalId(UUID.randomUUID())
+      .fields(fieldItems);
   }
 
-  private static FieldItem getFieldItem(String authorityId) {
-    return getFieldItem().authorityId(UUID.fromString(authorityId)).authorityNaturalId("12345");
-  }
-
-  private static FieldItem getFieldItemLinked() {
-    return getFieldItem(AUTHORITY_ID).authorityNaturalId("12345").linkingRuleId(LINKING_RULE_ID);
-  }
-
-  private static InstanceLink getInstanceLink(UUID authorityId) {
-    return new InstanceLink(RandomUtils.nextInt(),
-      authorityId,
-      "12345",
-      UUID.fromString("b9a5f035-de63-4e2c-92c2-07240c89b817"),
-      LINKING_RULE_ID);
-  }
-
-  private QuickMarc getQuickMarc(List<FieldItem> fieldItems) {
-    return new QuickMarc().marcFormat(MarcFormat.BIBLIOGRAPHIC)
+  private QuickMarcEdit getQuickMarcEdit(List<FieldItem> fieldItems) {
+    return new QuickMarcEdit().marcFormat(MarcFormat.BIBLIOGRAPHIC)
       .externalId(UUID.randomUUID())
       .fields(fieldItems);
   }