+++ src/test/java/org/folio/qm/service/population/impl/AuthorityLeaderMarcPopulationServiceTest.java
@@ -4,8 +4,8 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
+import org.folio.qm.domain.dto.BaseMarcRecord;
 import org.folio.qm.domain.dto.MarcFormat;
-import org.folio.qm.domain.dto.QuickMarc;
 import org.folio.qm.support.types.UnitTest;
 import org.junit.jupiter.api.Test;
 
@@ -79,8 +79,8 @@ class AuthorityLeaderMarcPopulationServiceTest {
     assertEquals(VALID_LEADER, quickMarc.getLeader());
   }
 
-  private QuickMarc getQuickMarc(String leader) {
-    return new QuickMarc()
+  private BaseMarcRecord getQuickMarc(String leader) {
+    return new BaseMarcRecord()
       .leader(leader);
   }
 }